Add Commitments to Offerings
This BG script loops throw all services and offerings in defined Service Portfolio and add Commitments selected in the array. The Names of these Commitments are first checked for existence. The script offers Dry Run capability to check how many updates will it perform. It can be easily tweaked to run it not on All services in portfolio, but for each service, when we need different commitments on some services/offerings.