Signature pad widget (#2472)
* Create README.md

Explains the features of the widget.

* Create HTML File.html

Defines the visual structure of the signature pad. It includes a <canvas> for drawing and two buttons for clearing or attaching the signature.

* Create Client Controller.js

Handles the logic for drawing on the canvas, capturing mouse/touch input, clearing the signature, and converting it into a Base64-encoded image for further processing.